fix(setupserver): fix body parsing
To parse HTTP request bodies on Node, one must accumulate incoming chunks into a Buffer. The body
parsing on `createContext` was written in such a way that these chunks would be accumulated into a
pre-existing Buffer initialized to 128 null bytes. It was my belief that the chunks would be written
*into* this Buffer, which would be resized by Node if it were to exceed 128 bytes. Instead, the
chunks are *concatenated to* the buffer. Therefore, all request bodies were being padded at their
start with 128 null bytes. The size of the initial Buffer was changed to zero, fixing this issue.

There was one test for setupServer, creating an echo server, which
should have caught this bug, but didn't. The assertion on the response body was
not strong enough, just checking that it matched the request body as a
regex, not that it was exactly equal to it. Because of that, it missed
the extra null bytes. Now the assertion was strenghtened to what it
should've been from the beginning.

BREAKING CHANGE: Since the body parsing was modified, code relying on the old behavior may stop
working.